+++ pkgs/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # pgflow
 
+## 0.2.6
+
+### Patch Changes
+
+- 7eed629: Make compile generate migration timestamp using UTC
+  - @pgflow/core@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- 69390b4: Make sure install command creates migrations using UTC timestamps
+  - @pgflow/core@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes