+++ pkgs/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pgflow/core
 
+## 0.2.6
+
+### Patch Changes
+
+- @pgflow/dsl@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- @pgflow/dsl@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes