+++ pkgs/example-flows/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pgflow/example-flows
 
+## 0.2.6
+
+### Patch Changes
+
+- @pgflow/core@0.2.6
+- @pgflow/dsl@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- @pgflow/core@0.2.5
+- @pgflow/dsl@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes